move Exception to own namespace, Process to own namespace
